fastApi/Blog/src/blog/token.py
from datetime import datetime, timedelta, timezone
import jwt

from . import schemas

SECRET_KEY = "09d25e094faa6ca2556c818166b7a9563b93f7099f6f0f4caa6cf63b88e8d3e7"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

# Create a JWT access token
def create_access_token(data: dict, expires_delta: timedelta | None = None):
    to_encode = data.copy()

    if expires_delta:
        expire = datetime.now(timezone.utc) + expires_delta
    else:
        expire = datetime.now(timezone.utc) + timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

# Verify and decode a JWT access token
def verify_access_token(token: str, credentials_exception):
    try:
            payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
            email = payload.get("sub")
            username = payload.get("username")
            user_id = payload.get("id")
            print(payload)
            if email is None:
                raise credentials_exception
            token_data = schemas.token_user(id=user_id,email=email,username=username)
            return token_data
    except jwt.InvalidTokenError:
        raise credentials_exception