fastApi/Blog/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.45
psycopg2-binary
pwdlib[argon2]
PyJWT

# Create environment:> python -m venv blog-env
# Activate environment:> .\blog-env\Scripts\Activate.ps1
# Install using requirements.txt file:> pip install -r .\requirements.txt
# Run Project :> uvicorn app.main:app --reload

